move messages to message.
add action methods.
