+++ src/main/java/tr/com/infumia/claimplugin/paper/api/permission/Permissions.java
@@ -6,12 +6,19 @@ import java.util.HashSet;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.function.Consumer;
 import org.jetbrains.annotations.NotNull;
+import tr.com.infumia.claimplugin.paper.api.event.LocationalEvent;
 
 /**
  * a class that contains default permissions for Claim plugin.
  */
 final class Permissions {
+
+  /**
+   * the actions.
+   */
+  private static final Map<ControlResult, Collection<Consumer<LocationalEvent>>> ACTIONS = new ConcurrentHashMap<>();
 
   /**
    * the permissions.
@@ -44,6 +51,18 @@ final class Permissions {
   @NotNull
   static Optional<Permission> get(@NotNull final String id) {
     return Optional.ofNullable(Permissions.PERMISSIONS.get(id));
+  }
+
+  /**
+   * gets action via result.
+   *
+   * @param result the result to get.
+   *
+   * @return action.
+   */
+  @NotNull
+  static Optional<Collection<Consumer<LocationalEvent>>> get(@NotNull final ControlResult result) {
+    return Optional.ofNullable(Permissions.ACTIONS.get(result));
   }
 
   /**
@@ -95,5 +114,17 @@ final class Permissions {
    */
   static void register(@NotNull final Permission permission) {
     Permissions.PERMISSIONS.put(permission.getId(), permission);
+  }
+
+  /**
+   * registers the actions.
+   *
+   * @param result the result to register.
+   * @param action the action to register.
+   */
+  static void register(@NotNull final ControlResult result, @NotNull final Consumer<LocationalEvent> action) {
+    final var actions = Permissions.ACTIONS.getOrDefault(result, new HashSet<>());
+    actions.add(action);
+    Permissions.ACTIONS.put(result, actions);
   }
 }
